CM 30-09-2021  - CMS Argumentos

+++ app/Http/Controllers/Intranet/Empresas/WikuController.php
@@ -4,8 +4,12 @@ namespace App\Http\Controllers\Intranet\Empresas;
 
 use App\Http\Controllers\Controller;
 use App\Models\Admin\WikuArea;
+use App\Models\Admin\WikuArgCriterio;
 use App\Models\Admin\WikuArgumento;
 use App\Models\Admin\WikuAsociacion;
+use App\Models\Admin\WikuAsociacionArg;
+use App\Models\Admin\WikuAutor;
+use App\Models\Admin\WikuAutorInst;
 use App\Models\Admin\WikuCriterio;
 use App\Models\Admin\WikuDocument;
 use App\Models\Admin\WikuNorma;
@@ -30,7 +34,7 @@ class WikuController extends Controller
     {
         $fuentes = WikuDocument::all();
         $normas = WikuNorma::all();
-        $argumentos = WikuArgumento::all();
+        $argumentos = WikuArgumento::with('autorInst', 'autor')->get();
         return view('intranet.parametros.wiku.index', compact('normas', 'fuentes', 'argumentos'));
     }
 
@@ -122,6 +126,32 @@ class WikuController extends Controller
                 $areas = WikuArea::all();
                 $temasEspecifico = WikuTemaEspecifico::all();
                 return view('intranet.parametros.wiku.normas.editar', compact('norma', 'fuentes', 'areas', 'temasEspecifico'));
+            }
+        } elseif ($wiku == 'argumento') {
+            if ($id == 0) {
+                $temasEspecifico = WikuTemaEspecifico::all();
+                $areas = WikuArea::all();
+                $autoresInst = WikuAutorInst::all();
+                $autores = WikuAutor::all();
+                return view('intranet.parametros.wiku.argumentos.crear', compact(
+                    'autoresInst',
+                    'autores',
+                    'temasEspecifico',
+                    'areas'
+                ));
+            } else {
+                $temasEspecifico = WikuTemaEspecifico::all();
+                $areas = WikuArea::all();
+                $autoresInst = WikuAutorInst::all();
+                $autores = WikuAutor::all();
+                $argumento = WikuArgumento::findOrFail($id);
+                return view('intranet.parametros.wiku.argumentos.editar', compact(
+                    'argumento',
+                    'autoresInst',
+                    'autores',
+                    'areas',
+                    'temasEspecifico'
+                ));
             }
         }
     }
@@ -262,6 +292,19 @@ class WikuController extends Controller
             $areas = WikuArea::all();
             $norma = WikuNorma::findOrFail($id);
             return view('intranet.parametros.wiku.normas.editar', compact('norma', 'fuentes', 'temasEspecifico', 'areas'));
+        } elseif ($wiku == 'argumento') {
+            $temasEspecifico = WikuTemaEspecifico::all();
+            $areas = WikuArea::all();
+            $argumento = WikuArgumento::findOrFail($id);
+            $autoresInst = WikuAutorInst::all();
+            $autores = WikuAutor::all();
+            return view('intranet.parametros.wiku.argumentos.editar', compact(
+                'argumento',
+                'temasEspecifico',
+                'areas',
+                'autoresInst',
+                'autores'
+            ));
         }
     }
     public function index_criterios($id, $wiku)
@@ -308,17 +351,34 @@ class WikuController extends Controller
             $areas = WikuArea::all();
             $norma = WikuNorma::findOrFail($id);
             return view('intranet.parametros.wiku.normas.editar', compact('norma', 'fuentes', 'temasEspecifico', 'areas'));
+        } elseif ($wiku == 'argumento') {
+            $temasEspecifico = WikuTemaEspecifico::all();
+            $areas = WikuArea::all();
+            $argumento = WikuArgumento::findOrFail($id);
+            $autoresInst = WikuAutorInst::all();
+            $autores = WikuAutor::all();
+
+
+            return view('intranet.parametros.wiku.argumentos.editar', compact(
+                'argumento',
+                'temasEspecifico',
+                'areas',
+                'autoresInst',
+                'autores'
+            ));
         }
     }
     public function index_palabras($id, $wiku)
     {
-        $norma = WikuNorma::findOrFail($id);
-        /*$palabras = WikuPalabras::with('normas')->whereHas('normas', function ($q) use ($id) {
-            $q->where('wiku_norma_id', $id);
-        })->get();*/
-        $palabras = WikuPalabras::all();
-
-        return view('intranet.parametros.wiku.palabras.index', compact('norma', 'palabras', 'wiku'));
+        if ($wiku == 'norma') {
+            $norma = WikuNorma::findOrFail($id);
+            $palabras = WikuPalabras::all();
+            return view('intranet.parametros.wiku.palabras.index', compact('norma', 'palabras', 'wiku'));
+        } elseif ($wiku == 'argumento') {
+            $argumento = WikuArgumento::findOrFail($id);
+            $palabras = WikuPalabras::all();
+            return view('intranet.parametros.wiku.palabras.index', compact('argumento', 'palabras', 'wiku'));
+        }
     }
     public function crear_palabras($id, $wiku)
     {
@@ -326,21 +386,37 @@ class WikuController extends Controller
     }
     public function guardar_palabras(Request $request, $id, $wiku)
     {
-
-        $palabras = WikuPalabras::with('normas')->whereHas('normas', function ($q) use ($id) {
-            $q->where('wiku_norma_id', $id);
-        })->get();
-        foreach ($palabras as $palabra) {
-            $palabras_norma[] = $palabra['id'];
+        if ($wiku == 'norma') {
+            $palabras = WikuPalabras::with('normas')->whereHas('normas', function ($q) use ($id) {
+                $q->where('wiku_norma_id', $id);
+            })->get();
+            foreach ($palabras as $palabra) {
+                $palabras_norma[] = $palabra['id'];
+            }
+            $nueva_palabra = $request->all();
+            unset($nueva_palabra['norma_id']);
+            $palabraNueva = WikuPalabras::create($nueva_palabra);
+            $palabras_norma[] = $palabraNueva->id;
+            $norma = WikuNorma::findOrFail($id);
+            $norma->palabras()->sync($palabras_norma);
+            $palabras = WikuPalabras::all();
+            return redirect('/admin/funcionario/wiku_palabras/index/' . $id . '/norma')->with('mensaje', 'Palabra creada con éxito y asociada a la norma')->with('norma')->with('palabras')->with('wiku');
+        } elseif ($wiku == 'argumento') {
+            $palabras = WikuPalabras::with('argumentos')->whereHas('argumentos', function ($q) use ($id) {
+                $q->where('wiku_argumento_id', $id);
+            })->get();
+            foreach ($palabras as $palabra) {
+                $palabras_argumento[] = $palabra['id'];
+            }
+            $nueva_palabra = $request->all();
+            unset($nueva_palabra['norma_id']);
+            $palabraNueva = WikuPalabras::create($nueva_palabra);
+            $palabras_argumento[] = $palabraNueva->id;
+            $argumento = WikuArgumento::findOrFail($id);
+            $argumento->palabras()->sync($palabras_argumento);
+            $palabras = WikuPalabras::all();
+            return redirect('/admin/funcionario/wiku_palabras/index/' . $id . '/argumento')->with('mensaje', 'Palabra creada con éxito y asociada al argumento')->with('argumento')->with('palabras')->with('wiku');
         }
-        $nueva_palabra = $request->all();
-        unset($nueva_palabra['norma_id']);
-        $palabraNueva = WikuPalabras::create($nueva_palabra);
-        $palabras_norma[] = $palabraNueva->id;
-        $norma = WikuNorma::findOrFail($id);
-        $norma->palabras()->sync($palabras_norma);
-        $palabras = WikuPalabras::all();
-        return redirect('/admin/funcionario/wiku_palabras/index/' . $id . '/norma')->with('mensaje', 'Palabra creada con éxito y asociada a la norma')->with('norma')->with('palabras')->with('wiku');
     }
     public function editar_palabras($id_palabras, $id, $wiku)
     {
@@ -352,15 +428,21 @@ class WikuController extends Controller
         $nueva_palabra = $request->all();
         unset($nueva_palabra['norma_id']);
         WikuPalabras::findOrFail($id_palabras)->update($nueva_palabra);
-        $norma = WikuNorma::findOrFail($id);
-        $palabras = WikuPalabras::all();
-        return redirect('/admin/funcionario/wiku_palabras/index/' . $id . '/norma')->with('mensaje', 'Palabra actualizada con éxito')->with('norma')->with('palabras')->with('wiku');
+        if ($wiku == 'norma') {
+            $norma = WikuNorma::findOrFail($id);
+            $palabras = WikuPalabras::all();
+            return redirect('/admin/funcionario/wiku_palabras/index/' . $id . '/norma')->with('mensaje', 'Palabra actualizada con éxito')->with('norma')->with('palabras')->with('wiku');
+        } elseif ($wiku == 'argumento') {
+            $argumento = WikuArgumento::findOrFail($id);
+            $palabras = WikuPalabras::all();
+            return redirect('/admin/funcionario/wiku_palabras/index/' . $id . '/argumento')->with('mensaje', 'Palabra actualizada con éxito')->with('argumento')->with('palabras')->with('wiku');
+        }
     }
     public function wiku_palabras_eliminar(Request $request, $id)
     {
         if ($request->ajax()) {
             $palabra = WikuPalabras::findOrfail($id);
-            if ($palabra->normas->count() == 0) {
+            if ($palabra->normas->count() == 0 && $palabra->argumentos->count() == 0) {
                 if (WikuPalabras::destroy($id)) {
                     return response()->json(['mensaje' => 'ok']);
                 } else {
@@ -373,30 +455,50 @@ class WikuController extends Controller
             abort(404);
         }
     }
-    public function adicionar_palabras(Request $request, $id_palabras, $id)
+    public function adicionar_palabras(Request $request, $id_palabras, $id, $wiku)
     {
         if ($request->ajax()) {
-            $palabra_ini = WikuPalabras::findOrfail($id_palabras);
-            $norma = WikuNorma::findOrFail($id);
-            $palabras = WikuPalabras::with('normas')->whereHas('normas', function ($q) use ($id) {
-                $q->where('wiku_norma_id', $id);
-            })->get();
-            foreach ($palabras as $palabra) {
-                $palabras_norma[] = $palabra['id'];
+            if ($wiku == 'norma') {
+                $palabra_ini = WikuPalabras::findOrfail($id_palabras);
+                $norma = WikuNorma::findOrFail($id);
+                $palabras = WikuPalabras::with('normas')->whereHas('normas', function ($q) use ($id) {
+                    $q->where('wiku_norma_id', $id);
+                })->get();
+                foreach ($palabras as $palabra) {
+                    $palabras_norma[] = $palabra['id'];
+                }
+                $palabras_norma[] = $palabra_ini->id;
+                $norma->palabras()->sync($palabras_norma);
+                return response()->json(['mensaje' => 'ok']);
+            } elseif ($wiku == 'argumento') {
+                $palabra_ini = WikuPalabras::findOrfail($id_palabras);
+                $argumento = WikuArgumento::findOrFail($id);
+                $palabras = WikuPalabras::with('argumentos')->whereHas('argumentos', function ($q) use ($id) {
+                    $q->where('wiku_argumento_id', $id);
+                })->get();
+                foreach ($palabras as $palabra) {
+                    $palabras_norma[] = $palabra['id'];
+                }
+                $palabras_norma[] = $palabra_ini->id;
+                $argumento->palabras()->sync($palabras_norma);
+                return response()->json(['mensaje' => 'ok']);
             }
-            $palabras_norma[] = $palabra_ini->id;
-            $norma->palabras()->sync($palabras_norma);
-            return response()->json(['mensaje' => 'ok']);
         } else {
             abort(404);
         }
     }
-    public function restar_palabras(Request $request, $id_palabras, $id)
+    public function restar_palabras(Request $request, $id_palabras, $id, $wiku)
     {
         if ($request->ajax()) {
-            $normas = new WikuNorma();
-            $normas->find($id)->palabras()->detach($id_palabras);
-            return response()->json(['mensaje' => 'ok']);
+            if ($wiku == 'norma') {
+                $normas = new WikuNorma();
+                $normas->find($id)->palabras()->detach($id_palabras);
+                return response()->json(['mensaje' => 'ok']);
+            } elseif ($wiku == 'argumento') {
+                $argumentos = new WikuArgumento();
+                $argumentos->find($id)->palabras()->detach($id_palabras);
+                return response()->json(['mensaje' => 'ok']);
+            }
         } else {
             abort(404);
         }
@@ -619,10 +721,18 @@ class WikuController extends Controller
     {
         $temasEspecifico = WikuTemaEspecifico::all();
         $areas = WikuArea::all();
-        return view('intranet.parametros.wiku.argumentos.crear', compact('temasEspecifico', 'areas'));
+        $autoresInst = WikuAutorInst::all();
+        $autores = WikuAutor::all();
+        return view('intranet.parametros.wiku.argumentos.crear', compact(
+            'temasEspecifico',
+            'areas',
+            'autoresInst',
+            'autores'
+        ));
     }
     public function guardar_argumento(Request $request)
     {
+        unset($request['autortipo']);
         WikuArgumento::create($request->all());
         return redirect('admin/funcionario/wiku/index')->with('mensaje', 'Argumento creado con exito.');
     }
@@ -631,15 +741,120 @@ class WikuController extends Controller
         $temasEspecifico = WikuTemaEspecifico::all();
         $areas = WikuArea::all();
         $argumento = WikuArgumento::findOrFail($id);
-        return view('intranet.parametros.wiku.argumentos.editar', compact('argumento', 'temasEspecifico', 'areas'));
+        $autoresInst = WikuAutorInst::all();
+        $autores = WikuAutor::all();
+        return view('intranet.parametros.wiku.argumentos.editar', compact(
+            'argumento',
+            'temasEspecifico',
+            'areas',
+            'autoresInst',
+            'autores'
+        ));
     }
     public function actualizar_argumento(Request $request, $id)
     {
+        if (is_null($request['publico'])) {
+            $request['publico'] = 0;
+        }
+        unset($request['autortipo']);
         WikuArgumento::findOrFail($id)->update($request->all());
         return redirect('admin/funcionario/wiku/index')->with('mensaje', 'Argumento actualizado con exito.');
     }
+    //------------------------------------------------------------------------------------------
+    public function cargarautori(Request $request)
+    {
+        if ($request->ajax()) {
 
+            WikuAutorInst::create($request->all());
+            return WikuAutorInst::get();
+        }
+    }
+    public function cargarautor(Request $request)
+    {
+        if ($request->ajax()) {
+
+            WikuAutor::create($request->all());
+            return WikuAutor::get();
+        }
+    }
     //************************************************************************************************
+    public function index_argcriterios($id, $wiku)
+    {
+        $argumento = WikuArgumento::findOrFail($id);
+        $criterios = WikuArgCriterio::where('argumento_id', $id)->get();
+        return view('intranet.parametros.wiku.argcriterios.index', compact('argumento', 'criterios', 'wiku'));
+    }
+    public function crear_argcriterios($id, $wiku)
+    {
+        return view('intranet.parametros.wiku.argcriterios.crear', compact('id', 'wiku'));
+    }
+    public function guardar_argcriterios(Request $request, $id, $wiku)
+    {
+        $nuevo_criterio = $request->all();
+        unset($nuevo_criterio['tipo_criterio']);
+        WikuArgCriterio::create($nuevo_criterio);
+        $argumento = WikuArgumento::findOrFail($id);
+        $criterios = WikuArgCriterio::where('argumento_id', $id)->get();
+        return redirect('/admin/funcionario/wiku_argcriterios/index/' . $id . '/argumento')->with('mensaje', 'Criterio creado con éxito')->with('argumento')->with('criterios')->with('wiku');
+    }
+    public function editar_argcriterios($id_criterios, $id, $wiku)
+    {
+        $criterio = WikuArgCriterio::findOrFail($id_criterios);
+        return view('intranet.parametros.wiku.argcriterios.editar', compact('criterio', 'id', 'wiku'));
+    }
+    public function actualizar_argcriterios(Request $request, $id_criterios, $id, $wiku)
+    {
+        $nuevo_criterio = $request->all();
+        unset($nuevo_criterio['tipo_criterio']);
+        WikuArgCriterio::findOrFail($id_criterios)->update($nuevo_criterio);
+        $argumento = WikuArgumento::findOrFail($id);
+        $criterios = WikuArgCriterio::where('argumento_id', $id)->get();
+        return redirect('/admin/funcionario/wiku_argcriterios/index/' . $id . '/argumento')->with('mensaje', 'Criterio actualizado con éxito')->with('argumento')->with('criterios')->with('wiku');
+    }
+    public function wiku_argcriterios_eliminar(Request $request, $id)
+    {
+        if ($request->ajax()) {
+            if (WikuArgCriterio::destroy($id)) {
+                return response()->json(['mensaje' => 'ok']);
+            } else {
+                return response()->json(['mensaje' => 'ng']);
+            }
+        } else {
+            abort(404);
+        }
+    }
     //================================================================================================
+    public function crear_argasociacion($id, $wiku)
+    {
+        $tipos_pqr = tipoPQR::get();
+        $categorias = Categoria::get();
+        $servicios = Servicio::get();
+        return view('intranet.parametros.wiku.argasociacion.crear', compact('id', 'wiku', 'tipos_pqr', 'categorias', 'servicios',));
+    }
+    public function guardar_argasociacion(Request $request, $id, $wiku)
+    {
+
+        if ($request['servicio_id'] == null) {
+            $request['prodserv'] = 'Producto';
+        } else {
+            $request['prodserv'] = 'Servicio';
+        }
+        $request['wiku_argumento_id'] = $id;
+        WikuAsociacionArg::create($request->all());
+        return redirect('/admin/funcionario/wiku/asociacion/crear/' . $id . '/argumento')->with('mensaje', 'Asociación actualizada con éxito')->with('argumento')->with('palabras')->with('wiku');
+    }
+    public function wiku_argasociacion_eliminar(Request $request, $id)
+    {
+        if ($request->ajax()) {
+            if (WikuAsociacionArg::destroy($id)) {
+                return response()->json(['mensaje' => 'ok']);
+            } else {
+                return response()->json(['mensaje' => 'ng']);
+            }
+        } else {
+            abort(404);
+        }
+    }
+    //------------------------------------------------------------------------------------------
 
 }
